Refactor medicine entity mapping and introduce detailed medicine information
- Updated `MedicineRepository` to query from the `vendor_medicine` table and commented out the unused `findBySubcategoryId` method.
- Introduced the `MedicineInformation` entity to store detailed metadata, such as descriptions, safety advice, interactions, and images, linked to a medicine via a one-to-one relationship.
- Changed the data type of `Stock` ID from `long` to `Integer`.
- Updated `PreOrderServiceImpl` to reflect the type change of `vendorId` from `long` to `Integer` during cart processing.

+++ src/main/java/com/trego/service/impl/PreOrderServiceImpl.java
@@ -310,7 +310,7 @@ public class PreOrderServiceImpl implements IPreOrderService {
                 .flatMap(cart -> cart.getMedicine().stream()
                         .map(medicine -> {
 
-                            long vendorId = cart.getVendorId();
+                            Integer vendorId = cart.getVendorId();
                             long medicineId = medicine.getId();
                             BigDecimal qty = BigDecimal.valueOf(medicine.getQty());
 
@@ -348,7 +348,7 @@ public class PreOrderServiceImpl implements IPreOrderService {
                 .flatMap(cart -> cart.getMedicine().stream()
                         .map(medicine -> {
 
-                            long vendorId = cart.getVendorId();
+                            Integer vendorId = cart.getVendorId();
                             long medicineId = medicine.getId();
                             int qty = medicine.getQty();
